requirements.txt
asgiref==3.6.0
Django==3.2.5
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.6
Pillow==9.5.0
django-widget-tweaks==1.4.12
django-crum==0.7.9
